pyproject.toml
[tool.poetry]
name = "flipr-api"
version = "1.2.0"
description = "Python client for flipr API."
authors = ["cnico"]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/cnico/flipr-api"
repository = "https://github.com/cnico/flipr-api"
documentation = "https://flipr-api.readthedocs.io"
keywords = ["flipr", "flipr-api", "pool", "piscine"]
classifiers = [
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: MIT License",
]

[tool.poetry.urls]
Changelog = "https://github.com/cnico/flipr-api/releases"

[tool.poetry.dependencies]
python = "^3.6.1"
requests = "^2.25.0"
python-dateutil = "^2.7.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
requests_mock = "^1.8.0"
coverage = {extras = ["toml"], version = "^5.3"}
safety = "^1.9.0"
mypy = "^0.790"
typeguard = "^2.10.0"
xdoctest = {extras = ["colors"], version = "^0.15.0"}
sphinx = "^3.3.1"
flake8 = "^3.8.4"
black = "^20.8b1"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^20.11.1"
flake8-docstrings = "^1.5.0"
flake8-rst-docstrings = "^0.0.14"
pep8-naming = "^0.11.1"
darglint = "^1.5.8"
reorder-python-imports = "^2.3.6"
sphinx-rtd-theme = "^0.5.0"
sphinx-click = "^2.5.0"
Pygments = "^2.7.3"

[tool.poetry.scripts]
flipr-api = "flipr_api.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["flipr_api"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
